Try to fix VC++ errors

+++ companion/src/eeprominterface.cpp
@@ -1693,10 +1693,11 @@ unsigned long LoadEeprom(RadioData &radioData, const uint8_t *eeprom, const int 
   std::bitset<NUM_ERRORS> errors;
 
   foreach(EEPROMInterface *eepromInterface, eepromInterfaces) {
-    std::bitset<NUM_ERRORS> result(eepromInterface->load(radioData, eeprom, size));
+    std::bitset<NUM_ERRORS> result((unsigned long long)eepromInterface->load(radioData, eeprom, size));
     if (result.test(NO_ERROR)) {
       return result.to_ulong();
-    } else {
+    }
+    else {
       errors |= result;
     }
   }
@@ -1707,15 +1708,16 @@ unsigned long LoadEeprom(RadioData &radioData, const uint8_t *eeprom, const int 
   return errors.to_ulong();
 }
 
-unsigned long LoadBackup(RadioData &radioData, uint8_t *eeprom, int size, int index)
+unsigned long LoadBackup(RadioData & radioData, uint8_t * eeprom, int size, int index)
 {
   std::bitset<NUM_ERRORS> errors;
 
   foreach(EEPROMInterface *eepromInterface, eepromInterfaces) {
-    std::bitset<NUM_ERRORS> result(eepromInterface->loadBackup(radioData, eeprom, size, index));
+    std::bitset<NUM_ERRORS> result((unsigned long long)eepromInterface->loadBackup(radioData, eeprom, size, index));
     if (result.test(NO_ERROR)) {
       return result.to_ulong();
-    } else {
+    }
+    else {
       errors |= result;
     }
   }
@@ -1727,15 +1729,16 @@ unsigned long LoadBackup(RadioData &radioData, uint8_t *eeprom, int size, int in
 }
 
 
-unsigned long LoadEepromXml(RadioData &radioData, QDomDocument &doc)
+unsigned long LoadEepromXml(RadioData & radioData, QDomDocument & doc)
 {
   std::bitset<NUM_ERRORS> errors;
 
   foreach(EEPROMInterface *eepromInterface, eepromInterfaces) {
-    std::bitset<NUM_ERRORS> result(eepromInterface->loadxml(radioData, doc));
+    std::bitset<NUM_ERRORS> result((unsigned long long)eepromInterface->loadxml(radioData, doc));
     if (result.test(NO_ERROR)) {
       return result.to_ulong();
-    } else {
+    }
+    else {
       errors |= result;
     }
   }
